backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
openai==1.30.0
langchain==0.2.1
langchain-openai==0.1.8
langchain-community==0.2.1
pydantic==2.7.1
python-multipart==0.0.9

# File readers
PyMuPDF==1.24.3
python-docx==1.1.2

# Web search
duckduckgo-search==6.1.0
google-search-results==2.4.2  # SerpAPI (optional)

# Utilities
aiofiles==23.2.1
httpx==0.27.0
